+++ src/hooks/game/index.ts
@@ -9,3 +9,6 @@ export type { BaseGameConfig, BaseGameState } from './use-base-game-state';
 
 export { useUIGameState } from './use-ui-game-state';
 export type { UIGameState } from './use-ui-game-state';
+
+export { useAdaptiveGuessGame } from '../use-adaptive-guess-game';
+export { useDecadeGameState } from '../use-decade-game-state';